shout_api/pyproject.toml
[tool.poetry]
name = "shout-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.14"
django = "^5.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
tweepy = "^4.14.0"
google-auth = "^2.33.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.140.0"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
